+++ src/platform_impl/linux/wayland/window/mod.rs
@@ -33,7 +33,7 @@ use super::event_loop::sink::EventSink;
 use super::output::MonitorHandle;
 use super::state::WinitState;
 use super::types::xdg_activation::XdgActivationTokenData;
-use super::{EventLoopWindowTarget, WaylandError, WindowId};
+use super::{ActiveEventLoop, WaylandError, WindowId};
 
 pub(crate) mod state;
 
@@ -81,7 +81,7 @@ pub struct Window {
 
 impl Window {
     pub(crate) fn new(
-        event_loop_window_target: &EventLoopWindowTarget,
+        event_loop_window_target: &ActiveEventLoop,
         attributes: WindowAttributes,
     ) -> Result<Self, RootOsError> {
         let queue_handle = event_loop_window_target.queue_handle.clone();