+++ tests/sync_object.rs
@@ -9,7 +9,7 @@ fn window_sync() {
 
 #[test]
 fn window_builder_sync() {
-    needs_sync::<winit::window::WindowBuilder>();
+    needs_sync::<winit::window::WindowAttributes>();
 }
 
 #[test]